Use root path to client buildInfo.json
The relative path does not work when the app is initialized in a subdirectory (e.g., `localhost/app/`). Then webpack is still bundling and serving the buildInfo.json from `localhost/buildInfo.json`. Hence, we must use the root path.

Alternative would be to change the webpack bundle and serving and make it flexible.